+++ src/Microsoft.Health.Dicom.Web/appsettings.json
@@ -85,7 +85,7 @@
     "MaxRetries": 3,
     "MinRetryDelayMilliseconds": 100,
     "Routes": {
-      "StatusTemplate": "Orchestrations/{0}"
+      "StatusTemplate": "Orchestrations/Instances/{0}"
     }
   }
 }
